suggests fields refactoring, clear form on multiedit dialog close, add suggestions to multiedit description field
