+++ src/containers/admin/PaymentCard/queries.js
@@ -1,5 +1,3 @@
-import pluralize from 'utils/pluralize'
-
 const PAYMENTS = `
 id: pid
 postedOn
@@ -26,62 +24,6 @@ categoryUpdater {
   isSystem
 }
 `
-
-const suggestDescriptions = [
-  `
-    query(
-      $accountId: ID!
-      $paymentId: ID!
-      $search: String
-    ) {
-      account(pid: $accountId) {
-        payment(pid: $paymentId) {
-          suggestedDescriptions(search: $search) {
-            text: description
-            count
-          }
-        }
-      }
-    }
-    `,
-  ({
-    account: {
-      payment: { suggestedDescriptions },
-    },
-  }) =>
-    suggestedDescriptions.map(({ text, count }) => ({
-      text,
-      data: text,
-      secondaryText: pluralize('payment', count),
-    })),
-]
-
-const suggestPeers = [
-  `
-  query(
-    $accountId: ID!
-    $search: String
-  ) {
-    account(pid: $accountId) {
-      peers(
-        sortBy: name_ASC
-        search: $search
-      ) {
-        id: pid
-        name
-        count: countPayments
-      }
-    }
-  }
-    `,
-  ({ account: { peers } }) =>
-    peers.map(({ id, name, count }) => ({
-      text: name,
-      data: { id, name },
-      secondaryText: pluralize('payment', count),
-    })),
-]
-
 const updatePayment = [
   `
     mutation(
@@ -121,7 +63,5 @@ const updatePayment = [
 ]
 
 export default {
-  suggestDescriptions,
-  suggestPeers,
   updatePayment,
 }
